feat: add smart mode to dbus and move codes
